requirements.txt
torch==2.7.1
torchac==0.9.3
torchvision==0.22.1
compressai==1.2.6
imagecodecs==2025.3.30
opencv_python==4.11.0.86
numpy==2.3.2
Pillow==11.3.0
prettytable==3.16.0
scipy==1.16.1
timm==1.0.19
accelerate
kornia
einops
huggingface_hub
tqdm
